+++ cvat-core/src/project-implementation.js
@@ -76,6 +76,16 @@
             return importDataset(this, format, file, updateStatusCallback);
         };
 
+        projectClass.prototype.backup.implementation = async function () {
+            const result = await serverProxy.projects.backupProject(this.id);
+            return result;
+        };
+
+        projectClass.restore.implementation = async function (file) {
+            const result = await serverProxy.projects.restoreProject(file);
+            return result.id;
+        };
+
         return projectClass;
     }