+++ cvat/apps/engine/tests/test_rest_api.py
@@ -16,6 +16,7 @@ from enum import Enum
 from glob import glob
 from io import BytesIO
 from unittest import mock
+import logging
 
 import av
 import numpy as np
@@ -34,6 +35,9 @@ from cvat.apps.engine.models import (AttributeSpec, AttributeType, Data, Job, Pr
     SortingMethod)
 from cvat.apps.engine.media_extractors import ValidateDimension, sort
 from utils.dataset_manifest import ImageManifestManager, VideoManifestManager
+
+#supress av warnings
+logging.getLogger('libav').setLevel(logging.ERROR)
 
 def create_db_users(cls):
     (group_admin, _) = Group.objects.get_or_create(name="admin")
@@ -1396,6 +1400,391 @@ class ProjectListOfTasksAPITestCase(APITestCase):
         response = self._run_api_v1_projects_id_tasks(None, project.id)
         self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
 
+class ProjectBackupAPITestCase(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        create_db_users(cls)
+        cls._create_media()
+        cls.client = APIClient()
+        cls._create_projects()
+
+    @classmethod
+    def tearDownClass(cls):
+        super().tearDownClass()
+        for task in cls.tasks:
+            shutil.rmtree(os.path.join(settings.TASKS_ROOT, str(task["id"])))
+            shutil.rmtree(os.path.join(settings.MEDIA_DATA_ROOT, str(task["data_id"])))
+
+        for f in cls.media['files']:
+            os.remove(f)
+        for d in cls.media['dirs']:
+            shutil.rmtree(d)
+
+    @classmethod
+    def _create_media(cls):
+        cls.media_data = []
+        cls.media = {
+            'files': [],
+            'dirs': [],
+        }
+        image_count = 10
+        imagename_pattern = "test_{}.jpg"
+        for i in range(image_count):
+            filename = imagename_pattern.format(i)
+            path = os.path.join(settings.SHARE_ROOT, filename)
+            cls.media['files'].append(path)
+            _, data = generate_image_file(filename)
+            with open(path, "wb") as image:
+                image.write(data.read())
+
+        cls.media_data.append(
+            {
+                **{"image_quality": 75,
+                   "copy_data": True,
+                   "start_frame": 2,
+                   "stop_frame": 9,
+                   "frame_filter": "step=2",
+                },
+                **{"server_files[{}]".format(i): imagename_pattern.format(i) for i in range(image_count)},
+            }
+        )
+
+        filename = "test_video_1.mp4"
+        path = os.path.join(settings.SHARE_ROOT, filename)
+        cls.media['files'].append(path)
+        _, data = generate_video_file(filename, width=1280, height=720)
+        with open(path, "wb") as video:
+            video.write(data.read())
+        cls.media_data.append(
+            {
+                "image_quality": 75,
+                "copy_data": True,
+                "start_frame": 2,
+                "stop_frame": 24,
+                "frame_filter": "step=2",
+                "server_files[0]": filename,
+            }
+        )
+
+        filename = os.path.join("test_archive_1.zip")
+        path = os.path.join(settings.SHARE_ROOT, filename)
+        cls.media['files'].append(path)
+        _, data = generate_zip_archive_file(filename, count=5)
+        with open(path, "wb") as zip_archive:
+            zip_archive.write(data.read())
+        cls.media_data.append(
+            {
+                "image_quality": 75,
+                "server_files[0]": filename,
+            }
+        )
+
+        filename = os.path.join("videos", "test_video_1.mp4")
+        path = os.path.join(settings.SHARE_ROOT, filename)
+        cls.media['dirs'].append(os.path.dirname(path))
+        os.makedirs(os.path.dirname(path))
+        _, data = generate_video_file(filename, width=1280, height=720)
+        with open(path, "wb") as video:
+            video.write(data.read())
+
+        manifest_path = os.path.join(settings.SHARE_ROOT, 'videos', 'manifest.jsonl')
+        generate_manifest_file(data_type='video', manifest_path=manifest_path, sources=[path])
+
+        cls.media_data.append(
+            {
+                "image_quality": 70,
+                "copy_data": True,
+                "server_files[0]": filename,
+                "server_files[1]": os.path.join("videos", "manifest.jsonl"),
+                "use_cache": True,
+            }
+        )
+
+        manifest_path = manifest_path=os.path.join(settings.SHARE_ROOT, 'manifest.jsonl')
+        generate_manifest_file(data_type='images', manifest_path=manifest_path,
+            sources=[os.path.join(settings.SHARE_ROOT, imagename_pattern.format(i)) for i in range(1, 8)])
+        cls.media['files'].append(manifest_path)
+        cls.media_data.append(
+            {
+                **{"image_quality": 70,
+                    "copy_data": True,
+                    "use_cache": True,
+                    "frame_filter": "step=2",
+                    "server_files[0]": "manifest.jsonl",
+                },
+                **{
+                    **{"server_files[{}]".format(i): imagename_pattern.format(i) for i in range(1, 8)},
+                }
+            }
+        )
+
+        cls.media_data.extend([
+            # image list local
+            {
+                "client_files[0]": generate_image_file("test_1.jpg")[1],
+                "client_files[1]": generate_image_file("test_2.jpg")[1],
+                "client_files[2]": generate_image_file("test_3.jpg")[1],
+                "image_quality": 75,
+            },
+            # video local
+            {
+                "client_files[0]": generate_video_file("test_video.mp4")[1],
+                "image_quality": 75,
+            },
+            # zip archive local
+            {
+                "client_files[0]": generate_zip_archive_file("test_archive_1.zip", 10)[1],
+                "image_quality": 50,
+            },
+            # pdf local
+            {
+                "client_files[0]": generate_pdf_file("test_pdf_1.pdf", 7)[1],
+                "image_quality": 54,
+            },
+        ])
+
+    @classmethod
+    def _create_tasks(cls, project):
+        def _create_task(task_data, media_data):
+            response = cls.client.post('/api/v1/tasks', data=task_data, format="json")
+            assert response.status_code == status.HTTP_201_CREATED
+            tid = response.data["id"]
+
+            for media in media_data.values():
+                if isinstance(media, io.BytesIO):
+                    media.seek(0)
+            response = cls.client.post("/api/v1/tasks/{}/data".format(tid), data=media_data)
+            assert response.status_code == status.HTTP_202_ACCEPTED
+            response = cls.client.get("/api/v1/tasks/{}".format(tid))
+            data_id = response.data["data"]
+            cls.tasks.append({
+                "id": tid,
+                "data_id": data_id,
+            })
+
+        task_data = [
+            {
+                "name": "my task #1",
+                "owner_id": cls.owner.id,
+                "assignee_id": cls.assignee.id,
+                "overlap": 0,
+                "segment_size": 100,
+                "project_id": project.id,
+            },
+            {
+                "name": "my task #2",
+                "owner_id": cls.owner.id,
+                "assignee_id": cls.assignee.id,
+                "overlap": 1,
+                "segment_size": 3,
+                "project_id": project.id,
+            },
+        ]
+
+        with ForceLogin(cls.owner, cls.client):
+            for data in task_data:
+                for media in cls.media_data:
+                    _create_task(data, media)
+
+    @classmethod
+    def _create_projects(cls):
+        cls.projects = []
+        cls.tasks = []
+        data = {
+            "name": "my empty project",
+            "owner": cls.owner,
+            "assignee": cls.assignee,
+            "labels": [{
+                "name": "car",
+                "color": "#ff00ff",
+                "attributes": [{
+                    "name": "bool_attribute",
+                    "mutable": True,
+                    "input_type": AttributeType.CHECKBOX,
+                    "default_value": "true"
+                }],
+                }, {
+                    "name": "person",
+                },
+            ],
+        }
+        db_project = create_db_project(data)
+        cls.projects.append(db_project)
+
+        data = {
+            "name": "my project without assignee",
+            "owner": cls.user,
+            "labels": [{
+                "name": "car",
+                "color": "#ff00ff",
+                "attributes": [{
+                    "name": "bool_attribute",
+                    "mutable": True,
+                    "input_type": AttributeType.CHECKBOX,
+                    "default_value": "true"
+                }],
+                }, {
+                    "name": "person",
+                },
+            ],
+        }
+        db_project = create_db_project(data)
+        cls._create_tasks(db_project)
+        cls.projects.append(db_project)
+
+        data = {
+            "name": "my big project",
+            "owner": cls.owner,
+            "assignee": cls.assignee,
+            "labels": [{
+                "name": "car",
+                "color": "#ff00ff",
+                "attributes": [{
+                    "name": "bool_attribute",
+                    "mutable": True,
+                    "input_type": AttributeType.CHECKBOX,
+                    "default_value": "true"
+                }],
+                }, {
+                    "name": "person",
+                },
+            ],
+        }
+        db_project = create_db_project(data)
+        cls._create_tasks(db_project)
+        cls.projects.append(db_project)
+
+        data = {
+            "name": "public project",
+            "labels": [{
+                "name": "car",
+                "color": "#ff00ff",
+                "attributes": [{
+                    "name": "bool_attribute",
+                    "mutable": True,
+                    "input_type": AttributeType.CHECKBOX,
+                    "default_value": "true"
+                }],
+                }, {
+                    "name": "person",
+                },
+            ],
+        }
+        db_project = create_db_project(data)
+        cls._create_tasks(db_project)
+        cls.projects.append(db_project)
+
+        data = {
+            "name": "super project",
+            "owner": cls.admin,
+            "assignee": cls.assignee,
+            "labels": [{
+                "name": "car",
+                "color": "#ff00ff",
+                "attributes": [{
+                    "name": "bool_attribute",
+                    "mutable": True,
+                    "input_type": AttributeType.CHECKBOX,
+                    "default_value": "true"
+                }],
+                }, {
+                    "name": "person",
+                },
+            ],
+        }
+        db_project = create_db_project(data)
+        cls._create_tasks(db_project)
+        cls.projects.append(db_project)
+
+    def _run_api_v1_projects_id_export(self, pid, user, query_params=""):
+        with ForceLogin(user, self.client):
+            response = self.client.get('/api/v1/projects/{}/backup?{}'.format(pid, query_params), format="json")
+
+        return response
+
+    def _run_api_v1_projects_import(self, user, data):
+        with ForceLogin(user, self.client):
+            response = self.client.post('/api/v1/projects/backup', data=data, format="multipart")
+
+        return response
+
+    def _run_api_v1_projects_id(self, pid, user):
+        with ForceLogin(user, self.client):
+            response = self.client.get('/api/v1/projects/{}'.format(pid), format="json")
+
+        return response.data
+
+    def _run_api_v1_projects_id_export_import(self, user):
+        for project in self.projects:
+            if user:
+                if user is self.user and (project.assignee or not project.owner):
+                    HTTP_200_OK = status.HTTP_403_FORBIDDEN
+                    HTTP_202_ACCEPTED = status.HTTP_403_FORBIDDEN
+                    HTTP_201_CREATED = status.HTTP_403_FORBIDDEN
+                else:
+                    HTTP_200_OK = status.HTTP_200_OK
+                    HTTP_202_ACCEPTED = status.HTTP_202_ACCEPTED
+                    HTTP_201_CREATED = status.HTTP_201_CREATED
+            else:
+                HTTP_200_OK = status.HTTP_401_UNAUTHORIZED
+                HTTP_202_ACCEPTED = status.HTTP_401_UNAUTHORIZED
+                HTTP_201_CREATED = status.HTTP_401_UNAUTHORIZED
+
+            pid = project.id
+            response = self._run_api_v1_projects_id_export(pid, user)
+            self.assertEqual(response.status_code, HTTP_202_ACCEPTED)
+
+            response = self._run_api_v1_projects_id_export(pid, user)
+            self.assertEqual(response.status_code, HTTP_201_CREATED)
+
+            response = self._run_api_v1_projects_id_export(pid, user, "action=download")
+            self.assertEqual(response.status_code, HTTP_200_OK)
+
+            if user and user is not self.observer and user is not self.user and user is not self.annotator:
+                self.assertTrue(response.streaming)
+                content = io.BytesIO(b"".join(response.streaming_content))
+                content.seek(0)
+
+                uploaded_data = {
+                    "project_file": content,
+                }
+                response = self._run_api_v1_projects_import(user, uploaded_data)
+                self.assertEqual(response.status_code, HTTP_202_ACCEPTED)
+                if user is not self.observer and user is not self.user and user is not self.annotator:
+                    rq_id = response.data["rq_id"]
+                    response = self._run_api_v1_projects_import(user, {"rq_id": rq_id})
+                    self.assertEqual(response.status_code, HTTP_201_CREATED)
+                    original_project = self._run_api_v1_projects_id(pid, user)
+                    imported_project = self._run_api_v1_projects_id(response.data["id"], user)
+                    compare_objects(
+                        self=self,
+                        obj1=original_project,
+                        obj2=imported_project,
+                        ignore_keys=(
+                            "data",
+                            "id",
+                            "url",
+                            "owner",
+                            "assignee",
+                            "created_date",
+                            "updated_date",
+                            "training_project",
+                            "project_id",
+                            "tasks",
+                        ),
+                    )
+
+    def test_api_v1_projects_id_export_admin(self):
+        self._run_api_v1_projects_id_export_import(self.admin)
+
+    def test_api_v1_projects_id_export_user(self):
+        self._run_api_v1_projects_id_export_import(self.user)
+
+    def test_api_v1_projects_id_export_observer(self):
+        self._run_api_v1_projects_id_export_import(self.observer)
+
+    def test_api_v1_projects_id_export_no_auth(self):
+        self._run_api_v1_projects_id_export_import(None)
 class ProjectExportAPITestCase(APITestCase):
     def setUp(self):
         self.client = APIClient()
@@ -1772,7 +2161,6 @@ class TaskDeleteAPITestCase(APITestCase):
         for task in self.tasks:
             task_dir = task.get_task_dirname()
             self.assertFalse(os.path.exists(task_dir))
-
 
 class TaskUpdateAPITestCase(APITestCase):
 
@@ -2305,8 +2693,6 @@ class TaskCreateAPITestCase(APITestCase):
         }
         self._check_api_v1_tasks(None, data)
 
-
-
 class TaskImportExportAPITestCase(APITestCase):
 
     def setUp(self):
@@ -2603,13 +2989,13 @@ class TaskImportExportAPITestCase(APITestCase):
 
     def _run_api_v1_tasks_id_export(self, tid, user, query_params=""):
         with ForceLogin(user, self.client):
-            response = self.client.get('/api/v1/tasks/{}?{}'.format(tid, query_params), format="json")
+            response = self.client.get('/api/v1/tasks/{}/backup?{}'.format(tid, query_params), format="json")
 
         return response
 
     def _run_api_v1_tasks_id_import(self, user, data):
         with ForceLogin(user, self.client):
-            response = self.client.post('/api/v1/tasks?action=import', data=data, format="multipart")
+            response = self.client.post('/api/v1/tasks/backup', data=data, format="multipart")
 
         return response
 
@@ -2637,10 +3023,10 @@ class TaskImportExportAPITestCase(APITestCase):
         self._create_tasks()
         for task in self.tasks:
             tid = task["id"]
-            response = self._run_api_v1_tasks_id_export(tid, user, "action=export")
+            response = self._run_api_v1_tasks_id_export(tid, user)
             self.assertEqual(response.status_code, HTTP_202_ACCEPTED)
 
-            response = self._run_api_v1_tasks_id_export(tid, user, "action=export")
+            response = self._run_api_v1_tasks_id_export(tid, user)
             self.assertEqual(response.status_code, HTTP_201_CREATED)
 
             response = self._run_api_v1_tasks_id_export(tid, user, "action=download")